+++ src/seahawk_deck/setup.py
@@ -21,6 +21,7 @@ setup(
     entry_points={
         'console_scripts': [
             "motion_controller=seahawk_deck.motion_controller:main",
+            "input_xbox_one=seahawk_deck.input_xbox_one:main",
             "rviz_markers=seahawk_deck.rviz_markers:main",
         ],
     },
